[Test] Remove test delegate indirection.
